+++ dash/dash.py
@@ -16,6 +16,7 @@ import hashlib
 import base64
 import traceback
 from urllib.parse import urlparse
+from typing import Union
 
 import flask
 
@@ -52,6 +53,7 @@ from ._utils import (
     to_json,
     convert_to_AttributeDict,
     gen_salt,
+    hooks_to_js_object,
 )
 from . import _callback
 from . import _get_paths
@@ -70,6 +72,7 @@ from ._pages import (
     _import_layouts_from_pages,
 )
 from ._jupyter import jupyter_dash, JupyterDisplayMode
+from .types import RendererHooks
 
 # Add explicit mapping for map files
 mimetypes.add_type("application/json", ".map", True)
@@ -134,7 +137,6 @@ except:  # noqa: E722
 
 
 def _get_traceback(secret, error: Exception):
-
     try:
         # pylint: disable=import-outside-toplevel
         from werkzeug.debug import tbtools
@@ -339,7 +341,13 @@ class Dash:
 
     :param add_log_handler: Automatically add a StreamHandler to the app logger
         if not added previously.
+
+    :param hooks: Extend Dash renderer functionality by passing a dictionary of
+    javascript functions. To hook into the layout, use dict keys "layout_pre" and
+    "layout_post". To hook into the callbacks, use keys "request_pre" and "request_post"
     """
+
+    _plotlyjs_url: str
 
     def __init__(  # pylint: disable=too-many-statements
         self,
@@ -373,6 +381,7 @@ class Dash:
         long_callback_manager=None,
         background_callback_manager=None,
         add_log_handler=True,
+        hooks: Union[RendererHooks, None] = None,
         **obsolete,
     ):
         _validate.check_obsolete(obsolete)
@@ -466,7 +475,7 @@ class Dash:
         self._favicon = None
 
         # default renderer string
-        self.renderer = "var renderer = new DashRenderer();"
+        self.renderer = f"var renderer = new DashRenderer({hooks_to_js_object(hooks)});"
 
         # static files from the packages
         self.css = Css(serve_locally)
@@ -588,6 +597,8 @@ class Dash:
         _get_app.APP = self
         self.enable_pages()
 
+        self._setup_plotlyjs()
+
     def _add_url(self, name, view_func, methods=("GET",)):
         full_name = self.config.routes_pathname_prefix + name
 
@@ -618,6 +629,25 @@ class Dash:
 
         # catch-all for front-end routes, used by dcc.Location
         self._add_url("<path:path>", self.index)
+
+    def _setup_plotlyjs(self):
+        # pylint: disable=import-outside-toplevel
+        from plotly.offline import get_plotlyjs_version
+
+        url = f"https://cdn.plot.ly/plotly-{get_plotlyjs_version()}.min.js"
+
+        # pylint: disable=protected-access
+        dcc._js_dist.extend(
+            [
+                {
+                    "relative_package_path": "package_data/plotly.min.js",
+                    "external_url": url,
+                    "namespace": "plotly",
+                    "async": "eager",
+                }
+            ]
+        )
+        self._plotlyjs_url = url
 
     @property
     def layout(self):
@@ -702,7 +732,10 @@ class Dash:
             "suppress_callback_exceptions": self.config.suppress_callback_exceptions,
             "update_title": self.config.update_title,
             "children_props": ComponentRegistry.children_props,
+            "serve_locally": self.config.serve_locally,
         }
+        if not self.config.serve_locally:
+            config["plotlyjs_url"] = self._plotlyjs_url
         if self._dev_tools.hot_reload:
             config["hot_reload"] = {
                 # convert from seconds to msec as used by js `setInterval`
@@ -1301,7 +1334,6 @@ class Dash:
 
         # Copy over global callback data structures assigned with `dash.callback`
         for k in list(_callback.GLOBAL_CALLBACK_MAP):
-
             if k in self.callback_map:
                 raise DuplicateCallback(
                     f"The callback `{k}` provided with `dash.callback` was already "
@@ -1328,7 +1360,6 @@ class Dash:
 
         if cancels:
             for cancel_input, manager in cancels.items():
-
                 # pylint: disable=cell-var-from-loop
                 @self.callback(
                     Output(cancel_input.component_id, "id"),
@@ -1719,7 +1750,6 @@ class Dash:
             _reload.watch_thread.start()
 
         if debug:
-
             if jupyter_dash.active:
                 jupyter_dash.configure_callback_exception_handling(
                     self, dev_tools.prune_errors
@@ -1753,7 +1783,6 @@ class Dash:
                     dash_total["dur"] = round((time.time() - dash_total["dur"]) * 1000)
 
                 for name, info in timing_information.items():
-
                     value = name
                     if info.get("desc") is not None:
                         value += f';desc="{info["desc"]}"'